Address reviewer comment
8, 3-7: This paragraph should be rewritten.  Please expand the idea of
why this method is useful - Revised paragraph to read, “The loose
coupling approach that we adopted for this research has proven to be
useful. It allows for us to examine patterns and trends that previously
we were unable to easily examine. Previously the RICEPEST model was not
able to be used to map yield losses, it was useful for comparing
production situtions only. The EPIRICE model could only be used to
estimate disease severity, not yield losses. This approach enhances the
usefulness of both models and should prove useful for further work
where spatial modelling of rice yield loss due to rice diseases is
desired.”